Make Physics work
Move the balls by their velocity each frame and otherwise do all the
little things necessary to essentially complete the Physics engine. I
still need to normalize the speeds based on the number of ms elapsed
since the last frame, but it is now possible to watch a simulation of
pool balls breaking.